new implementation of RPC required with new model (PhysicalCardWithEmailDealing), new return type and flow.
