+++ src/main/java/com/aadm/cardexchange/shared/ExchangeServiceAsync.java
@@ -7,6 +7,4 @@ import java.util.List;
 
 public interface ExchangeServiceAsync {
     void addProposal(String token, String receiverUserEmail, List<PhysicalCard> senderPhysicalCards, List<PhysicalCard> receiverPhysicalCards, AsyncCallback<Boolean> async);
-
-    void CheckExistingPcardByIdCard(String token, Game game, int cardId, AsyncCallback<Boolean> async);
 }
