+++ src/test/java/com/aadm/cardexchange/server/DeckServiceTest.java
@@ -12,6 +12,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
+import org.junit.jupiter.params.provider.NullSource;
 import org.junit.jupiter.params.provider.ValueSource;
 import org.mapdb.Serializer;
 
@@ -20,6 +21,7 @@ import javax.servlet.ServletContext;
 import javax.servlet.ServletException;
 import java.util.*;
 
+import static com.aadm.cardexchange.server.MockCardData.createMagicDummyMap;
 import static org.easymock.EasyMock.*;
 
 public class DeckServiceTest {
@@ -368,6 +370,24 @@ public class DeckServiceTest {
                 .andReturn(cardMap);
     }
 
+    private  Map<Integer, MagicCard>  generateValidMagicCardMap(){
+        return createMagicDummyMap();
+    }
+
+    private Map<String, Deck>  generateValidDeckofMagicPhysicalCardMap(Map<Integer, MagicCard> magicCardMap, int i, String deckName){
+        Map<String, Deck> deckMap = new HashMap<>() {{
+            put(deckName, new Deck(deckName, true));
+        }};
+        int count = 0;
+        for (Map.Entry<Integer, MagicCard> entry : magicCardMap.entrySet()) {
+            if(count<i) {
+                deckMap.get(deckName).addPhysicalCard(new PhysicalCard(Game.MAGIC, entry.getKey(), Status.randomStatus(), "This is a super valid description"));
+            }
+            count++;
+        }
+        return deckMap;
+    }
+
     @Test
     public void testGetDeckByNameForValidEntry() throws AuthException {
         setupForValidToken();
@@ -616,4 +636,179 @@ public class DeckServiceTest {
         Assertions.assertTrue(deckService.removeCustomDeck("validToken", deckName));
         ctrl.verify();
     }
+    @ParameterizedTest
+    @NullAndEmptySource
+    @ValueSource(strings = {"invalidToken"})
+    public void testCheckExistingPcardByCardForInvalidToken(String input) {
+        Map<String, LoginInfo> loginInfoMap = new HashMap<>() {{
+            put("validToken1", new LoginInfo("test@test.it", System.currentTimeMillis() - 10000));
+            put("validToken2", new LoginInfo("test2@test.it", System.currentTimeMillis() - 20000));
+            put("validToken3", new LoginInfo("test3@test.it", System.currentTimeMillis() - 30000));
+        }};
+        expect(mockConfig.getServletContext()).andReturn(mockCtx);
+        expect(mockDB.getPersistentMap(isA(ServletContext.class), anyString(), isA(Serializer.class), isA(Serializer.class)))
+                .andReturn(loginInfoMap);
+        ctrl.replay();
+        Assertions.assertThrows(AuthException.class, () -> deckService.getListPhysicalCardWithEmailDealing(input, Game.randomGame(), 1));
+        ctrl.verify();
+    }
+
+    @ParameterizedTest
+    @NullSource
+    public void testCheckExistingPcardByCardForNullGame(Game input) {
+        setupForValidToken();
+        ctrl.replay();
+        Assertions.assertThrows(InputException.class, () -> {
+            deckService.getListPhysicalCardWithEmailDealing("validToken", input, 1);
+            ctrl.verify();
+        });
+    }
+
+    @Test
+    public void testCheckExistingPcardByCardForNegativeCardId() {
+        setupForValidToken();
+        ctrl.replay();
+        Assertions.assertThrows(InputException.class, () -> {
+            deckService.getListPhysicalCardWithEmailDealing("validToken", Game.randomGame(), -1);
+            ctrl.verify();
+        });
+    }
+
+
+    @Test
+    public void testCheckExistingPcardByCardForUserWithoutDecksMap() {
+        setupForValidToken();
+        Map<String, Map<String, Deck>> mockDeckMap = new HashMap<>();
+        expect(mockConfig.getServletContext()).andReturn(mockCtx);
+        expect(mockDB.getPersistentMap(isA(ServletContext.class), anyString(), isA(Serializer.class), isA(Serializer.class)))
+                .andReturn(mockDeckMap);
+        ctrl.replay();
+        Assertions.assertThrows(RuntimeException.class, () ->
+                deckService.getListPhysicalCardWithEmailDealing("validToken", Game.randomGame(), 3)
+        );
+        ctrl.verify();
+    }
+
+
+    @Test
+    public void testCheckExistingPcardByCardForUserWithDeckNull() {
+        setupForValidToken();
+        Map<String, Map<String, Deck>> mockDeckMap = new HashMap<>() {{
+            put("test@test.it", null);
+        }};
+        expect(mockConfig.getServletContext()).andReturn(mockCtx);
+        expect(mockDB.getPersistentMap(isA(ServletContext.class), anyString(), isA(Serializer.class), isA(Serializer.class)))
+                .andReturn(mockDeckMap);
+        ctrl.replay();
+        Assertions.assertThrows(RuntimeException.class, () ->
+                deckService.getListPhysicalCardWithEmailDealing("validToken", Game.randomGame(), 3)
+        );
+        ctrl.verify();
+    }
+
+
+    @Test
+    public void testCheckExistingPcardByCardForUserWithEmptyDeckMap() {
+        setupForValidToken();
+        Map<String, Deck> deckMap = new HashMap<>();
+        Map<String, Map<String, Deck>> mockDeckMap = new HashMap<>() {{
+            put("test@test.it", deckMap);
+        }};
+        expect(mockConfig.getServletContext()).andReturn(mockCtx);
+        expect(mockDB.getPersistentMap(isA(ServletContext.class), anyString(), isA(Serializer.class), isA(Serializer.class)))
+                .andReturn(mockDeckMap);
+        ctrl.replay();
+        Assertions.assertThrows(BaseException.class, () ->
+                deckService.getListPhysicalCardWithEmailDealing("validToken", Game.randomGame(), 3)
+        );
+        ctrl.verify();
+    }
+
+    @Test
+    public void testCheckExistingPcardByCardWithoutOwnedDeck() throws InputException, AuthException {
+        //serve per catturare l'eccezione BaseException, con mazzo "Owned" non trovato
+        setupForValidToken();
+        Map<String, Deck> deckMap =  generateValidDeckofMagicPhysicalCardMap(generateValidMagicCardMap(),5,"owd");
+        Map<String, Map<String, Deck>> mockDeckMap = new HashMap<>() {{
+            put("test@test.it", deckMap);
+        }};
+        expect(mockConfig.getServletContext()).andReturn(mockCtx);
+        expect(mockDB.getPersistentMap(isA(ServletContext.class), anyString(), isA(Serializer.class), isA(Serializer.class)))
+                .andReturn(mockDeckMap);
+        ctrl.replay();
+        Assertions.assertThrows(BaseException.class, () ->
+                deckService.getListPhysicalCardWithEmailDealing("validToken", Game.randomGame(), 3)
+        );
+        ctrl.verify();
+    }
+
+                /*
+    @Test
+    public void testCheckExistingPcardByCardForUserWithoutCardsInOwnedDeck() throws BaseException {
+        setupForValidToken();
+        Map<String, Deck> deckMap = new HashMap<>() {{
+            put("Owned", new Deck("Wished", true));
+        }};
+        Map<String, Map<String, Deck>> mockDeckMap = new HashMap<>() {{
+            put("test@test.it", deckMap);
+        }};
+        expect(mockConfig.getServletContext()).andReturn(mockCtx);
+        expect(mockDB.getPersistentMap(isA(ServletContext.class), anyString(), isA(Serializer.class), isA(Serializer.class)))
+                .andReturn(mockDeckMap);
+        ctrl.replay();
+        Assertions.assertFalse(exchangeService.CheckExistingPcardByIdCard("validToken", Game.randomGame(), 3));
+        ctrl.verify();
+    }
+
+
+
+
+    @Test
+    public void testCheckExistingPcardByCardForUserWithoutPCardsInOwnedDeck() throws BaseException {
+        setupForValidToken();
+        Map<String, Deck> deckMap = new HashMap<>() {{
+            put("Owned", new Deck("Owned", true));
+        }};
+        Map<String, Map<String, Deck>> mockDeckMap = new HashMap<>() {{
+            put("test@test.it", deckMap);
+        }};
+        expect(mockConfig.getServletContext()).andReturn(mockCtx);
+        expect(mockDB.getPersistentMap(isA(ServletContext.class), anyString(), isA(Serializer.class), isA(Serializer.class)))
+                .andReturn(mockDeckMap);
+        ctrl.replay();
+        Assertions.assertFalse(exchangeService.CheckExistingPcardByIdCard("validToken", Game.randomGame(), 3));
+        ctrl.verify();
+    }
+
+    @Test
+    public void testCheckExistingPcardByCardNotMatchingIdCard() throws BaseException {
+        setupForValidToken();
+        Map<String, Deck> deckMap =  generateValidDeckofMagicPhysicalCardMap(generateValidMagicCardMap(),5,"Owned");
+        Map<String, Map<String, Deck>> mockDeckMap = new HashMap<>() {{
+            put("test@test.it", deckMap);
+        }};
+        expect(mockConfig.getServletContext()).andReturn(mockCtx);
+        expect(mockDB.getPersistentMap(isA(ServletContext.class), anyString(), isA(Serializer.class), isA(Serializer.class)))
+                .andReturn(mockDeckMap);
+        ctrl.replay();
+        Assertions.assertFalse(exchangeService.CheckExistingPcardByIdCard("validToken", Game.randomGame(), 55));
+        ctrl.verify();
+    }
+    @Test
+    public void testCheckExistingPcardByCardMatchingIdCard() throws BaseException {
+        setupForValidToken();
+        Map<String, Deck> deckMap =  generateValidDeckofMagicPhysicalCardMap(generateValidMagicCardMap(),5,"Owned");
+        Map<String, Map<String, Deck>> mockDeckMap = new HashMap<>() {{
+            put("test@test.it", deckMap);
+        }};
+        Set<PhysicalCard> PcardList = deckMap.get("Owned").getPhysicalCards();
+        int randomRealCardIdToFound = PcardList.stream().skip(new Random().nextInt(PcardList.size())).findFirst().orElse(null).getCardId();
+        expect(mockConfig.getServletContext()).andReturn(mockCtx);
+        expect(mockDB.getPersistentMap(isA(ServletContext.class), anyString(), isA(Serializer.class), isA(Serializer.class)))
+                .andReturn(mockDeckMap);
+        ctrl.replay();
+        Assertions.assertTrue(exchangeService.CheckExistingPcardByIdCard("validToken", Game.randomGame(), randomRealCardIdToFound));
+        ctrl.verify();
+    }
+*/
 }
